requirements.txt
psycopg2-binary
langchain-community
ollama
streamlit
python-dotenv
sqlalchemy
